Produccion v2.0

+++ app/Http/Controllers/Admin/MedicamentoController.php
@@ -5,8 +5,8 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Cache;
 use App\Models\Medicamento;
+use App\Models\User;
 use App\Http\Requests\MedicamentoRequest;
-use App\Models\TipoMedicamento;
 use Illuminate\Http\Request;
 
 class MedicamentoController extends Controller
@@ -22,15 +22,15 @@ class MedicamentoController extends Controller
     
     public function index()
     {
+        /* $this->authorize('author', $medicamento); */
         $medicamentos = Medicamento::all();
         return view('admin.medicamento.index', compact('medicamentos'));
     }
 
      public function create()
     {
-        $tipo_medicamento = TipoMedicamento::pluck('descripcion', 'id');
-        
-        return view('admin.medicamento.create', compact('tipo_medicamento'));
+        $users = User::pluck('name', 'id');
+        return view('admin.medicamento.create', compact('users'));
     }
 
     public function store(MedicamentoRequest $request)
@@ -46,13 +46,17 @@ class MedicamentoController extends Controller
     
     public function edit(Medicamento $medicamento )
     {
-        $tipo_medicamento = TipoMedicamento::pluck('descripcion', 'id');
-        return view('admin.medicamento.edit' , compact('medicamento', 'tipo_medicamento'));
+        $this->authorize('author', $medicamento);
+        
+        $users = User::pluck('name', 'id');
+        return view('admin.medicamento.edit' , compact('medicamento', 'users'));
    
     }
 
      public function update(MedicamentoRequest $request, Medicamento $medicamento)
     {
+        $this->authorize('author', $medicamento);
+       
         $medicamento->update($request->all());
         
         Cache::flush();
@@ -64,6 +68,8 @@ class MedicamentoController extends Controller
     
     public function destroy(Medicamento $medicamento)
     {
+        $this->authorize('author', $medicamento);
+        
         $medicamento->delete();
 
         Cache::flush();
